move DEFAULT_PRESET to constants to avoid circular import, update CLI to call create_agent_from_preset properly
