+++ memgpt/autogen/examples/agent_groupchat.py
@@ -13,8 +13,7 @@ Begin by doing:
 import os
 import autogen
 from memgpt.autogen.memgpt_agent import create_memgpt_autogen_agent_from_config
-from memgpt.presets.presets import DEFAULT_PRESET
-from memgpt.constants import LLM_MAX_TOKENS
+from memgpt.constants import LLM_MAX_TOKENS, DEFAULT_PRESET
 
 LLM_BACKEND = "openai"
 # LLM_BACKEND = "azure"
